+++ proofs/src/poly/kzg/msm.rs
@@ -1,5 +1,6 @@
 use std::fmt::Debug;
 
+use ff::Field;
 use group::{prime::PrimeCurveAffine, Curve, Group};
 use halo2curves::{
     msm::msm_best,
@@ -29,22 +30,18 @@ pub struct MSMKZG<E: Engine> {
 
 impl<E: Engine> MSMKZG<E> {
     /// Create an empty MSM instance
-    pub fn new() -> Self {
+    pub fn init() -> Self {
         MSMKZG {
             scalars: vec![],
             bases: vec![],
         }
     }
 
-    /// Prepares all scalars in the MSM to linear combination
-    pub fn combine_with_base(&mut self, base: E::Fr) {
-        use ff::Field;
-        let mut acc = E::Fr::ONE;
-        if !self.scalars.is_empty() {
-            for scalar in self.scalars.iter_mut().rev() {
-                *scalar *= &acc;
-                acc *= base;
-            }
+    /// Create a new MSM from a given base (with scalar of 1).
+    pub fn from_base(base: &E::G1) -> Self {
+        MSMKZG {
+            scalars: vec![E::Fr::ONE],
+            bases: vec![*base],
         }
     }
 }
@@ -110,7 +107,7 @@ where
     E::G1Affine: CurveAffine<ScalarExt = E::Fr, CurveExt = E::G1>,
 {
     fn default() -> Self {
-        Self::new()
+        Self::init()
     }
 }
 
@@ -131,15 +128,20 @@ impl<E: MultiMillerLoop + Debug> DualMSM<E>
 where
     E::G1Affine: CurveAffine<ScalarExt = E::Fr, CurveExt = E::G1>,
 {
-    /// Create a new two channel MSM accumulator instance
-    pub fn new() -> Self {
+    /// Create an empty two channel MSM accumulator instance
+    pub fn init() -> Self {
         Self {
-            left: MSMKZG::new(),
-            right: MSMKZG::new(),
+            left: MSMKZG::init(),
+            right: MSMKZG::init(),
         }
     }
 
-    /// Split the [DualMSM] into `left` and `right`.
+    /// Create a new two channel MSM accumulator instance
+    pub fn new(left: MSMKZG<E>, right: MSMKZG<E>) -> Self {
+        Self { left, right }
+    }
+
+    /// Split the [DualMSM] into `left` and `right`
     pub fn split(&self) -> SplitDualMSM<E> {
         let left = self
             .left